+++ src/main/java/com/streisky/virtualshowcasebackend/config/validation/exception/ResourceNotFoundException.java
@@ -1,0 +1,8 @@
+package com.streisky.virtualshowcasebackend.config.validation.exception;
+
+public class ResourceNotFoundException extends RuntimeException {
+
+    public ResourceNotFoundException(Long id) {
+        super("No record was found with the ID " + id);
+    }
+}
