+++ src/main/java/com/streisky/virtualshowcasebackend/domain/account/repository/AccountRepository.java
@@ -1,8 +1,8 @@
-package com.streisky.virtualshowcasebackend.repository.account;
+package com.streisky.virtualshowcasebackend.domain.account.repository;
 
 import java.util.Optional;
 
-import com.streisky.virtualshowcasebackend.entity.account.Account;
+import com.streisky.virtualshowcasebackend.domain.account.entity.Account;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface AccountRepository extends JpaRepository<Account, Long> {
