+++ src/main/java/com/streisky/virtualshowcasebackend/domain/product/repository/ProductRepository.java
@@ -1,8 +1,8 @@
-package com.streisky.virtualshowcasebackend.repository.product;
+package com.streisky.virtualshowcasebackend.domain.product.repository;
 
 import java.util.Optional;
 
-import com.streisky.virtualshowcasebackend.entity.product.Product;
+import com.streisky.virtualshowcasebackend.domain.product.entity.Product;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
