--- src/main/java/com/streisky/virtualshowcasebackend/exception/VirtualShowcaseNotFoundException.java
@@ -1,8 +1,0 @@
-package com.streisky.virtualshowcasebackend.exception;
-
-public class VirtualShowcaseNotFoundException extends RuntimeException {
-
-    public VirtualShowcaseNotFoundException(Long id) {
-        super("No record was found with the ID " + id);
-    }
-}
